requirements.txt
pyPDF2
streamlit
numpy
google-generativeai==0.8.3  